+++ lib/ts_packet.h
@@ -255,7 +255,7 @@ namespace TS{
 
   size_t getUniqTrackID(const DTSC::Meta &M, size_t idx);
 
-  const char *createPMT(std::set<unsigned long> &selectedTracks, const DTSC::Meta &M, int contCounter = 0);
+  const char *createPMT(std::set<size_t> &selectedTracks, const DTSC::Meta &M, int contCounter = 0);
   const char *createSDT(const std::string &streamName, int contCounter = 0);
 
 }// namespace TS
